+++ src/gql/ElementQueryConditionBuilder.php
@@ -74,6 +74,12 @@ class ElementQueryConditionBuilder extends Component
      * @var ResolveInfo
      */
     private $_resolveInfo;
+
+    /**
+     * @var ArgumentManager
+     */
+    private $_argumentManager;
+
     private $_fragments;
     private $_eagerLoadableFieldsByContext = [];
     private $_transformableAssetProperties = ['url', 'width', 'height'];
@@ -113,6 +119,17 @@ class ElementQueryConditionBuilder extends Component
     {
         $this->_resolveInfo = $resolveInfo;
         $this->_fragments = $this->_resolveInfo->fragments;
+    }
+
+    /**
+     * Set the current ResolveInfo object.
+     *
+     * @param ArgumentManager $argumentManager
+     * @since 3.6.0
+     */
+    public function setArgumentManager(ArgumentManager $argumentManager)
+    {
+        $this->_argumentManager = $argumentManager;
     }
 
     /**
@@ -191,7 +208,7 @@ class ElementQueryConditionBuilder extends Component
                     $extractedValue = $argumentNodeValue->value;
             }
         } else {
-            $extractedValue = $argumentNodeValue;
+            $extractedValue = $argumentNode->kind === 'IntValue' ? (int) $argumentNodeValue : $argumentNodeValue;
         }
 
         return $extractedValue;
@@ -453,7 +470,7 @@ class ElementQueryConditionBuilder extends Component
 
                         // For relational fields, prepare the arguments.
                         if ($craftContentField instanceof BaseRelationField) {
-                            $arguments = ElementResolver::prepareArguments($arguments);
+                            $arguments = $this->_argumentManager->prepareArguments($arguments);
                         }
                     }